LayoutTests/imported/w3c/web-platform-tests/service-workers/cache-storage/window/cache-put.https-expected.txt


PASS Cache.put called with simple Request and Response 
FAIL Cache.put called with Request and Response from fetch() assert_equals: Cache.put should store response body. expected "a simple text file\n" but got ""
PASS Cache.put with Request without a body 
PASS Cache.put with Response without a body 
FAIL Cache.put with a Response containing an empty URL assert_equals: Cache.put should update the cache with new Request and Response. Attributes differ: status. expected 200 but got 0
FAIL Cache.put with an empty response body assert_equals: Cache.put should store status. expected 200 but got 0
PASS Cache.put with synthetic 206 response 
PASS Cache.put with HTTP 206 response 
PASS Cache.put with HTTP 500 response 
FAIL Cache.put called twice with matching Requests and different Responses assert_equals: Cache.put should replace existing response with new response. Attributes differ: status. expected 200 but got 0
FAIL Cache.put called twice with request URLs that differ only by a fragment assert_equals: Cache.put should replace existing response with new response. Attributes differ: status. expected 200 but got 0
FAIL Cache.put with a string request assert_equals: Cache.put should accept a string as request. expected "some body" but got ""
PASS Cache.put with an invalid response 
PASS Cache.put with a non-HTTP/HTTPS request 
FAIL Cache.put with a relative URL assert_equals: Cache.put should accept a relative URL as the request. Attributes differ: status. expected 200 but got 0
PASS Cache.put with a non-GET request 
PASS Cache.put with a null response 
PASS Cache.put with a POST request 
PASS Cache.put with a used response body 
FAIL getReader() after Cache.put assert_throws: function "() => response.body.getReader()" did not throw
PASS Cache.put with a VARY:* Response 
PASS Cache.put with an embedded VARY:* Response 
FAIL Cache.put should store Response.redirect() correctly assert_equals: Redirect response is reproduced by the Cache API Attributes differ: status. expected 302 but got 0
